wk2108 | rework on review comments
