--- frame/collective/src/mock.rs
@@ -1,174 +1,0 @@
-// This file is part of Substrate.
-
-// Copyright (C) 2018-2021 Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: Apache-2.0
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//! Test utilities
-
-#![cfg(test)]
-
-use sp_std::{prelude::*};
-
-use sp_runtime::{
-	traits::{Hash, BlakeTwo256, IdentityLookup}, testing::Header,
-	BuildStorage,
-};
-use sp_core::H256;
-
-use frame_support::{
-	Hashable, assert_ok, assert_noop, parameter_types,
-	traits::{ChangeMembers},
-	weights::{Pays, GetDispatchInfo},
-};
-
-use frame_system::{self as system, EventRecord, Phase};
-use hex_literal::hex;
-use crate::{
-	self as collective,
-	Config, decl_tests,
-};
-parameter_types! {
-	pub const BlockHashCount: u64 = 250;
-	pub BlockWeights: frame_system::limits::BlockWeights =
-		frame_system::limits::BlockWeights::simple_max(1024);
-}
-impl frame_system::Config for Test {
-	type BaseCallFilter = ();
-	type BlockWeights = ();
-	type BlockLength = ();
-	type DbWeight = ();
-	type Origin = Origin;
-	type Index = u64;
-	type BlockNumber = u64;
-	type Call = Call;
-	type Hash = H256;
-	type Hashing = BlakeTwo256;
-	type AccountId = u64;
-	type Lookup = IdentityLookup<Self::AccountId>;
-	type Header = Header;
-	type Event = Event;
-	type BlockHashCount = BlockHashCount;
-	type Version = ();
-	type PalletInfo = PalletInfo;
-	type AccountData = ();
-	type OnNewAccount = ();
-	type OnKilledAccount = ();
-	type SystemWeightInfo = ();
-	type SS58Prefix = ();
-}
-
-/// Set the prime member's vote as the default vote.
-pub struct PrimeDefaultVote;
-
-impl collective::DefaultVote for PrimeDefaultVote {
-	fn default_vote(
-		prime_vote: Option<bool>,
-		_yes_votes: collective::MemberCount,
-		_no_votes: collective::MemberCount,
-		_len: collective::MemberCount,
-	) -> bool {
-		prime_vote.unwrap_or(false)
-	}
-}
-
-/// First see if yes vote are over majority of the whole collective. If so, set the default vote
-/// as yes. Otherwise, use the prime meber's vote as the default vote.
-pub struct MoreThanMajorityThenPrimeDefaultVote;
-
-impl collective::DefaultVote for MoreThanMajorityThenPrimeDefaultVote {
-	fn default_vote(
-		prime_vote: Option<bool>,
-		yes_votes: collective::MemberCount,
-		_no_votes: collective::MemberCount,
-		len: collective::MemberCount,
-	) -> bool {
-		let more_than_majority = yes_votes * 2 > len;
-		more_than_majority || prime_vote.unwrap_or(false)
-	}
-}
-
-parameter_types! {
-	pub const MotionDuration: u64 = 3;
-	pub const MaxProposals: u32 = 100;
-	pub const MaxMembers: u32 = 100;
-}
-impl Config<collective::Instance1> for Test {
-	type Origin = Origin;
-	type Proposal = Call;
-	type Event = Event;
-	type MotionDuration = MotionDuration;
-	type MaxProposals = MaxProposals;
-	type MaxMembers = MaxMembers;
-	type DefaultVote = PrimeDefaultVote;
-	type WeightInfo = ();
-}
-impl Config<collective::Instance2> for Test {
-	type Origin = Origin;
-	type Proposal = Call;
-	type Event = Event;
-	type MotionDuration = MotionDuration;
-	type MaxProposals = MaxProposals;
-	type MaxMembers = MaxMembers;
-	type DefaultVote = MoreThanMajorityThenPrimeDefaultVote;
-	type WeightInfo = ();
-}
-impl collective::Config for Test {
-	type Origin = Origin;
-	type Proposal = Call;
-	type Event = Event;
-	type MotionDuration = MotionDuration;
-	type MaxProposals = MaxProposals;
-	type MaxMembers = MaxMembers;
-	type DefaultVote = PrimeDefaultVote;
-	type WeightInfo = ();
-}
-
-type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-type Block = frame_system::mocking::MockBlock<Test>;
-
-frame_support::construct_runtime!(
-	pub enum Test where
-		Block = Block,
-		NodeBlock = Block,
-		UncheckedExtrinsic = UncheckedExtrinsic
-	{
-		System: system::{Module, Call, Event<T>},
-		Collective: collective::<Instance1>::{Module, Call, Storage, Event<T>, Origin<T>, Config<T>},
-		CollectiveMajority: collective::<Instance2>::{Module, Call, Storage, Event<T>, Origin<T>, Config<T>},
-		DefaultCollective: collective::{Module, Call, Storage, Event<T>, Origin<T>, Config<T>},
-	}
-);
-
-pub fn new_test_ext() -> sp_io::TestExternalities {
-	let mut ext: sp_io::TestExternalities = GenesisConfig {
-		collective_Instance1: Some(collective::GenesisConfig {
-			members: vec![1, 2, 3],
-			phantom: Default::default(),
-		}),
-		collective_Instance2: Some(collective::GenesisConfig {
-			members: vec![1, 2, 3, 4, 5],
-			phantom: Default::default(),
-		}),
-		collective: None,
-	}.build_storage().unwrap().into();
-	ext.execute_with(|| System::set_block_number(1));
-	ext
-}
-
-fn make_proposal(value: u64) -> Call {
-	Call::System(frame_system::Call::remark(value.encode()))
-}
-
-decl_tests!{ Test }
